requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.350
neo4j==5.14.1
pydantic==2.5.2
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.1
networkx==3.2.1
matplotlib==3.8.2
pytest==7.4.3
httpx==0.25.2